+++ share/isso.conf
@@ -18,10 +18,13 @@ name =
 # your website, you have to "whitelist" your website(s).
 #
 # I recommend the first value to be a non-SSL website that is used as fallback
-# if Firefox users (and only those) supress their HTTP referer completely.
+# if Firefox users (and only those) supress their HTTP referer completely:
+#
+# host =
+#     http://example.tld/
+#     https://example.tld/
+#
 host =
-    http://example.tld/
-    https://example.tld/
 
 # time range that allows users to edit/remove their own comments.
 # It supports years, weeks, days, hours, minutes, seconds.
@@ -38,7 +41,7 @@ max-age = 15m
 # smtp
 #     Send notifications via SMTP on new comments with activation (if
 #     moderated) and deletion links.
-notify =
+notify = stdout
 
 
 [moderation]
